Add the main functionality for execution and adjust project documentation.
The main function now includes a basic try/catch block to start the application in the most basic form.
The README has been adjusted to correct the branch/versioning format to use SemVer conventions.
Furthermore the gitignore has been adjusted to exclude .o files but include "opcpp_ic.o" as this is the API given and required to succesfully compile the project.